requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
google-images-download==2.8.0
h5py==2.10.0
idna==2.9
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
urllib3==1.25.8
